+++ src/Chinchilla.Specifications/Messages/TestRequestMessage.cs
@@ -2,7 +2,7 @@
 
 namespace Chinchilla.Specifications.Messages
 {
-    public class TestRequestMessage : ICorrelated
+    public class TestRequestMessage : ICorrelated, IHasTimeOut
     {
         public TestRequestMessage()
         {
@@ -10,5 +10,10 @@ namespace Chinchilla.Specifications.Messages
         }
 
         public Guid CorrelationId { get; set; }
+
+        public TimeSpan Timeout
+        {
+            get { return TimeSpan.FromSeconds(60); }
+        }
     }
 }